Split all client requests into seperate files
