+++ pumpkin-protocol/src/client/config/c_config_disconnect.rs
@@ -1,0 +1,16 @@
+use crate::bytebuf::packet_id::Packet;
+
+#[derive(serde::Serialize)]
+pub struct CConfigDisconnect<'a> {
+    reason: &'a str,
+}
+
+impl<'a> Packet for CConfigDisconnect<'a> {
+    const PACKET_ID: i32 = 0x02;
+}
+
+impl<'a> CConfigDisconnect<'a> {
+    pub fn new(reason: &'a str) -> Self {
+        Self { reason }
+    }
+}
