+++ pumpkin-protocol/src/client/status/mod.rs
@@ -1,38 +1,5 @@
-use crate::{
-    bytebuf::{packet_id::Packet, ByteBuffer},
-    ClientPacket, VarInt,
-};
+mod c_ping_response;
+mod c_status_response;
 
-#[derive(serde::Serialize)]
-pub struct CPingResponse {
-    payload: i64, // must responde with the same as in `SPingRequest`
-}
-
-impl Packet for CPingResponse {
-    const PACKET_ID: VarInt = 0x01;
-}
-
-impl CPingResponse {
-    pub fn new(payload: i64) -> Self {
-        Self { payload }
-    }
-}
-
-pub struct CStatusResponse<'a> {
-    json_response: &'a str, // 32767
-}
-
-impl<'a> CStatusResponse<'a> {
-    pub fn new(json_response: &'a str) -> Self {
-        Self { json_response }
-    }
-}
-impl<'a> Packet for CStatusResponse<'a> {
-    const PACKET_ID: VarInt = 0x00;
-}
-
-impl<'a> ClientPacket for CStatusResponse<'a> {
-    fn write(&self, bytebuf: &mut ByteBuffer) {
-        bytebuf.put_string(self.json_response);
-    }
-}
+pub use c_ping_response::*;
+pub use c_status_response::*;
